意图cluster产品/intention_cluster_gemini/requirements.txt
google-generativeai>=0.3.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.5.0
seaborn>=0.12.0

